QPdfEnginePrivate - fix for '-qreal float' build.
Literals 1. or 0. have type double, but QGradientStop.first is qreal and
thus can be float depending on platform/configuration, making
qBound call invalid (qBound(0., qreal_value, 1.) for example).

Task-number: QTBUG-47319

+++ src/gui/painting/qpdf.cpp
@@ -1979,8 +1979,8 @@ int QPdfEnginePrivate::createShadingFunction(const QGradient *gradient, int from
         if (reflect && step % 2) {
             for (int i = stops.size() - 1; i > 0; --i) {
                 QGradientBound b;
-                b.start = step + 1 - qBound(0., stops.at(i).first, 1.);
-                b.stop = step + 1 - qBound(0., stops.at(i - 1).first, 1.);
+                b.start = step + 1 - qBound(qreal(0.), stops.at(i).first, qreal(1.));
+                b.stop = step + 1 - qBound(qreal(0.), stops.at(i - 1).first, qreal(1.));
                 b.function = functions.at(i - 1);
                 b.reverse = true;
                 gradientBounds << b;
@@ -1988,8 +1988,8 @@ int QPdfEnginePrivate::createShadingFunction(const QGradient *gradient, int from
         } else {
             for (int i = 0; i < stops.size() - 1; ++i) {
                 QGradientBound b;
-                b.start = step + qBound(0., stops.at(i).first, 1.);
-                b.stop = step + qBound(0., stops.at(i + 1).first, 1.);
+                b.start = step + qBound(qreal(0.), stops.at(i).first, qreal(1.));
+                b.stop = step + qBound(qreal(0.), stops.at(i + 1).first, qreal(1.));
                 b.function = functions.at(i);
                 b.reverse = false;
                 gradientBounds << b;
